WindowServer+LibGfx: Improve non-standard bitmap pitch handling
Up until now, the only way to create a Bitmap with a non-standard
pitch was to use the wrapper; This was unsuitable for back and temp
buffers within WindowServer.

This commit adds another optional parameter for specifying the pitch
in a backwards-compatible way. The change in WindowServer will likely
fix most cases of screen misalignment when running on baremetal.