SpiceAgent: Make message handling much more reliable
Previously, the spice agent would attempt to read all chunks for a
message as soon as the spice chardev became readable. This generally was
fine for small messages but unreliable for large transfers such as
files. If at any point a read failed, the entire message would be
abandoned, then when notified that there's more data, the agent would
try to read it as a new chunk/message. If the data was the rest of a
file transfer it would end up parsing that as a chunk header, and
crashing when it tried to allocate a random amount of bytes.

This is resolved by making the agent only read what's available and
keeping track of how much data it's read. Then only once there's a
complete message does it attempt to handle it.