LibPDF: Check DecodeParms array element type before casting
0000337.pdf contains:

    /DecodeParms[(null)<</K -1/Columns 2545>>]

where the array's first entry `(null)` is a string with
the characters `"null"`. Casting the string to a dict used to crash.
Instead, treat non-dict array entries as if no decode parms were present
for the filter at that array position.

This will fix a crash for 0000337.pdf once we can decode /JPXDecode
filters.