LibGfx/JPEG2000: Add an implementation of the jpeg2000 IDWT algorithm
Only used in tests for now.

This is also written to look close to the spec.

The API is fairly allocation- and copy-heavy. We can improve this later
once everything works and we have solid tests.

Making this testable requires having a few structs that are similar
to some structs internal to JPEG2000Loader.

Also move `Transformation` to the new header file.