LibGfx/JPEG2000: Implement conversion of samples to Bitmap
This is fairly ad-hoc for now, but with this `image` can read many
JPEG2000 files and save them in another format -- jpeg2000 decoding
now works end-to-end :^)

It's not yet hooked up in LibPDF, it doesn't yet handle somewhat
more exotic color formats well, it misses support for all kinds of
relatively rare other JPEG2000 features, but the basics are working.

This also doesn't do any of the clever things that are possible with
jpeg2000, such as only decoding image components that are needed,
only decoding as many bits per channel as needed, only decoding
as much resolution as needed, or decoding a subrect of the image.
Maybe later.