Tests/LibGfx: Replace test image for #25678 with better one
The test image added in #25678 had wonky borders, due to
uclouvain/openjpeg#1575. Recreate it using jasper instead of
openjpeg, which doesn't have this problem.

jasper does have the problem that it doesn't write a cdef box,
which I think is invalid, so I built jasper with this local hack
to work around that:
https://github.com/jasper-software/jasper/issues/392#issuecomment-2631428226

Image created using

    build/src/app/jasper --input \
        Tests/LibGfx/test-inputs/jpeg2000/kakadu-lossless-*.jp2 \
        --output jasper-tile3x2-res5.jp2 \
        -O tilewidth=39 -O tileheight=53 -O numrlvls=5

I verified that the new image still triggers the crash without the
code change in #25678.