LibGfx/JPEG2000: Reject images with weird channel order
To make the test image, I opened ref.png, swapped the
red and green channels in Photoshop, saved it as
ref-bgr.bmp, then ran:

    ../openjpeg/build/bin/opj_compress -i ref-bgr.bmp \
        -o openjpeg-lossless-bgra-u8.jp2 \
        -mct 0 \
        -C "opj_compress -mct 0"

and then manually edited the file in a hex editor to
associate channel 0 with B (3) and channel 2 with R (1).
In other words, the byte at offset 0x5c changed from 0x01 to 0x03,
and the byte at offset 0x68 changed from 0x03 to 0x01.