Tests: Add more JPEG2000 bypass code block style test images
Adds one test for the interaction of bypass code block style
(which jasper calls "lazy") with termall code block style, created with:

    jasper --input \
        Tests/LibGfx/test-inputs/jpeg2000/jasper-tile3x2-res5.jp2 \
        --output .../jasper-rgba-u8-cbstyle-05-bypass-termall.jp2 \
        -O lazy -O termall

Adds one test for the interaction of bypass code block style with
layers, created with:

    jasper --input \
        Tests/LibGfx/test-inputs/jpeg2000/jasper-tile3x2-res5.jp2 \
        --output .../jasper-rgba-u8-cbstyle-01-bypass-layers.jp2 \
        -O lazy -O ilyrrates=0.1,0.5,1

Adds one test where the segment with the first 10 passes is spread out
over 3 layers, so that the middle layer contains a single segment that
continues the first segment from the first layer, and is further
continued in the third layer. The second layer contains not a single
complete segment. Created with:

    jasper --input \
        Tests/LibGfx/test-inputs/jpeg2000/jasper-tile3x2-res5.jp2 \
        --output .../jasper-rgba-u8-cbstyle-01-bypass-finer-layers.jp2 \
        -O lazy -O ilyrrates=0.002,0.009,1

Adds one test for the interaction of bypass code block style with
other code block styles, except the termall style.  (The "termination"
style mostly disables the "bypass" style, so having both on does not
really test how bypass interacts with other modes.) Created with:
interaction testing of "bypass".)

     jasper --input ... --output ... \
         -O lazy -O segsym -O vcausal -O pterm -O resetprob

None of these decode successfully yet, but they're useful for making
sure eventual support for the lazy code block style is somewhat
correct.