LibGfx/JPEG2000: Add segment_index_from_pass_index() function and use it
In particular, use it to compute the number of segments in a packet as
the difference of the last segment index in the current packet minues
the last segment index in the previous packet, possibly incremented by
one if the two packets share a segment (such as in every normal
file that does not have any code block styles set).

This is a bit more code, but it will elegantly handle the selective
arithmetic decoding bypass code block style as well.

No behavior change.