LibPDF: Read axial shading dict entries
We now create an AxialShading object and read all required
parameters and build objects for them, but we don't use them
for any rendering yet.

Mostly to still produce a "not yet implemented" diag for this,
give `Shading` a `draw()` method, call it from `Renderer`, and
return a "not yet implemented" error from `AxialShading::draw()`.