Kernel/ACPI: Always use the bootloader-provided RSDP address, if present
The BSP setup code and MADT finding code always tried to find the RSDP
in BIOS-specific memory areas. This was missed in 1197f1faab5.

Rename `find_rsdp_in_{platform => ia_pc}_specific_memory_location` and
make that function x86-only, since the RSDP can only be found in the
UEFI Configuration Table both on AArch64 and RISC-V.

The new `find_rsdp()` function tries to first get the RSDP from the
`BootInfo` struct (the EFIPrekernel stores the RSDP address from the
configuration table there). If `acpi_rsdp_paddr` is null, it tries to
find the RSDP in the BIOS areas on x86.