Toolchain: Update LLVM to 20.1.0
Recently, LLVM refactored the way in which locale support is
implemented, which means that the old way of stubbing all that out no
longer works, as the base implementation now expects far greater locale
support than what we have. To work around that, we now use the same
built-in implementations for a few locale-related functions as z/OS
does, which appears to work with fairly minimal changes.