LibGfx/JPEG2000: Add support for parsing J2K_RGN markers
We don't do anything with the data yet.

No behavior change, except that RGN marker segment contents are now
logged when building with `JPEG2000_DEBUG` defined to 1.

Also add a test image with a RGN marker, generated with:

    ../openjpeg/build/bin/opj_compress \
        -i ref.bmp -o output0-roi.jp2 \
        -ROI c=0,U=5 \
        -C 'opj_compress -ROI c=0,U=5'

If I just remove the error out for RGN markers, this seems to decode
fine without any changes, despite all of Annex H in T.800 being about
what to do if a RGN marker is present. Maybe that's because the
exponent or colors have to be set just right for this to have an
effect in lossless files? But currently the image is only used to
check that we still refuse to decode images with RGN markers, and it's
good enough for that. If we add actual decoding support, we'll likely
need a better test input.