DisplaySettings: Support disabling the desktop wallpaper
This adds a "None" wallpaper option to allow removing the desktop
wallpaper. This also fixes a longstanding bug that when opening
DisplaySettings without a wallpaper set the first wallpaper would
automatically be selected, marking the settings as modified without the
user having made any changes. (This was due to the model index being
invalid so the first index was automatically selected.)